+++ src/HeboTech.ATLib.Modems/Cinterion/MC55i.cs
@@ -1,7 +1,7 @@
 ﻿using HeboTech.ATLib.Modems.Generic;
 using HeboTech.ATLib.Parsing;
 using HeboTech.ATLib.Messaging;
-using HeboTech.ATLib.Sys;
+using HeboTech.ATLib.Misc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
