+++ src/HeboTech.ATLib.Modems/Generic/ModemBase.cs
@@ -4,7 +4,7 @@ using HeboTech.ATLib.Messaging;
 using HeboTech.ATLib.Numbering;
 using HeboTech.ATLib.Parsing;
 using HeboTech.ATLib.Storage;
-using HeboTech.ATLib.Sys;
+using HeboTech.ATLib.Misc;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
