+++ src/HeboTech.ATLib.Modems/SIMCOM/ISIM5320.cs
@@ -1,5 +1,5 @@
 ﻿using HeboTech.ATLib.Parsing;
-using HeboTech.ATLib.Sys;
+using HeboTech.ATLib.Misc;
 using System.Threading.Tasks;
 
 namespace HeboTech.ATLib.Modems.SIMCOM
