+++ src/HeboTech.ATLib.Modems/SIMCOM/SIM5320.cs
@@ -2,7 +2,7 @@
 using HeboTech.ATLib.Modems.Generic;
 using HeboTech.ATLib.Parsing;
 using HeboTech.ATLib.Messaging;
-using HeboTech.ATLib.Sys;
+using HeboTech.ATLib.Misc;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
