+++ src/HeboTech.ATLib.TestConsole/StressTest.cs
@@ -1,7 +1,7 @@
 ﻿using HeboTech.ATLib.Messaging;
 using HeboTech.ATLib.Modems.D_LINK;
 using HeboTech.ATLib.Parsing;
-using HeboTech.ATLib.Sys;
+using HeboTech.ATLib.Misc;
 using System;
 using System.Threading;
 using System.Threading.Tasks;
