+++ src/HeboTech.ATLib.Tests/Misc/ImsiTests.cs
@@ -1,7 +1,7 @@
-﻿using HeboTech.ATLib.Sys;
+﻿using HeboTech.ATLib.Misc;
 using Xunit;
 
-namespace HeboTech.ATLib.Tests.Sys
+namespace HeboTech.ATLib.Tests.Misc
 {
     public class ImsiTests
     {
