+++ src/HeboTech.ATLib/IModem.cs
@@ -3,7 +3,7 @@ using HeboTech.ATLib.Messaging;
 using HeboTech.ATLib.Numbering;
 using HeboTech.ATLib.Parsing;
 using HeboTech.ATLib.Storage;
-using HeboTech.ATLib.Sys;
+using HeboTech.ATLib.Misc;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
