+++ src/HeboTech.ATLib/Misc/BatteryStatus.cs
@@ -1,7 +1,7 @@
 ﻿using HeboTech.ATLib.Extensions;
 using UnitsNet;
 
-namespace HeboTech.ATLib.Sys
+namespace HeboTech.ATLib.Misc
 {
     public class BatteryStatus
     {
