+++ src/HeboTech.ATLib/Misc/Imsi.cs
@@ -1,4 +1,4 @@
-﻿namespace HeboTech.ATLib.Sys
+﻿namespace HeboTech.ATLib.Misc
 {
     public class Imsi
     {
