+++ src/HeboTech.ATLib/Misc/ProductIdentificationInformation.cs
@@ -1,4 +1,4 @@
-﻿namespace HeboTech.ATLib.Sys
+﻿namespace HeboTech.ATLib.Misc
 {
     public class ProductIdentificationInformation
     {
