+++ src/HeboTech.ATLib/Misc/RemainingPinPukAttempts.cs
@@ -1,4 +1,4 @@
-﻿namespace HeboTech.ATLib.Sys
+﻿namespace HeboTech.ATLib.Misc
 {
     public class RemainingPinPukAttempts
     {
